server/modules/load_vectorstore.py
import os
import time
from pathlib import Path
from dotenv import load_dotenv
from tqdm.auto import tqdm
from pinecone import Pinecone, ServerlessSpec
from langchain_community.document_loaders import PyPDFLoader
from langchain_text_splitters.character import RecursiveCharacterTextSplitter
from langchain_google_genai import GoogleGenerativeAIEmbeddings

load_dotenv()

GOOGLE_API_KEY=os.getenv("GOOGLE_API_KEY")
PINECONE_API_KEY=os.getenv("PINECONE_API_KEY")
PINECONE_ENV=os.getenv("PINECONE_ENV", "us-east-1")
PINECONE_INDEX_NAME=os.getenv("PINECONE_INDEX_NAME", "medical-index-v2")

os.environ["GOOGLE_API_KEY"]=GOOGLE_API_KEY

UPLOAD_DIR="./uploaded_docs"
os.makedirs(UPLOAD_DIR, exist_ok=True)


def load_vectorstore(uploaded_files):
    pc=Pinecone(api_key=PINECONE_API_KEY)
    spec=ServerlessSpec(cloud="aws", region=PINECONE_ENV)
    existing_indexes=[i["name"] for i in pc.list_indexes()]

    if PINECONE_INDEX_NAME not in existing_indexes:
        pc.create_index(name=PINECONE_INDEX_NAME, dimension=3072, metric="cosine", spec=spec)
        print(f"Creating Pinecone index '{PINECONE_INDEX_NAME}'...")
        while not pc.describe_index(PINECONE_INDEX_NAME).status["ready"]:
            print("Waiting for Pinecone index to be ready...")
            time.sleep(1)

    index=pc.Index(PINECONE_INDEX_NAME)
    
    embed_model=GoogleGenerativeAIEmbeddings(model="models/gemini-embedding-001")
    file_paths=[]

    for file in uploaded_files:
        save_path=Path(UPLOAD_DIR)/file.filename
        with open(save_path, "wb") as f:
            f.write(file.file.read())
        file_paths.append(str(save_path))

    for file_path in file_paths:
        loader=PyPDFLoader(file_path)
        documents=loader.load()

        splitter=RecursiveCharacterTextSplitter(chunk_size=500, chunk_overlap=100)
        chunks=splitter.split_documents(documents)

        texts=[chunk.page_content for chunk in chunks]
        metadata=[chunk.metadata for chunk in chunks]
        ids=[f"{Path(file_path).stem}-{i}" for i in range(len(chunks))]

        print(f"Embedding chunks from {file_path}...")
        embedding=embed_model.embed_documents(texts)

        print(f"Upserting embedding from {file_path} to Pinecone...")
        with tqdm(total=len(embedding), desc=f"Upserting to PineCone") as progress:
            index.upsert(vectors=zip(ids, embedding, metadata))
            progress.update(len(embedding))

        print(f"Finished processing {file_path}.")

        
